+++ 10-Toolchain-actions/run_03-Run-action-and-exit.py
@@ -9,7 +9,7 @@ import shutil
 import sys
 import tct
 
-VERSION = "v3.1.0"
+VERSION = "v3.2.0"
 
 params = tct.readjson(sys.argv[1])
 binabspath = sys.argv[2]